net: lib: Align upstream changes to mqtt_keepalive_time_left()
 - mqtt_keepalive_time_left() now returns -1 if keep alive messages are
   disabled. Handle this in the various cloud backends.

 - Use the various library wrappers for mqtt_keepalive_time_left() as
   the timeout argument in poll().

 - Add check upon a poll POLLIN event. If the client is disconnected
   break out of poll() to avoid waiting the remaining timeout.

 - Update state handling in azure_iot_hub and nrf_cloud libraries to
   properly handle the internal state with the new changes.

Disabling keepalive messages will cause mqtt_keepalive_time_lef() to
return -1. By sending in -1 for the timeout in poll(), poll will wait
forever.

[CIA-162]

+++ subsys/net/lib/nrf_cloud/src/nrf_cloud_transport.c
@@ -1128,7 +1128,7 @@ void nct_process(void)
 
 int nct_keepalive_time_left(void)
 {
-	return (int)mqtt_keepalive_time_left(&nct.client);
+	return mqtt_keepalive_time_left(&nct.client);
 }
 
 int nct_socket_get(void)
